+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.85.3] - 2020-11-13
+
 ### Fixed
 - Initial search when `queryArgs` is `undefined`.